+++ src/test/java/org/jabref/logic/exporter/BibtexDatabaseWriterTest.java
@@ -60,7 +60,6 @@ import static org.mockito.Mockito.when;
 
 class BibtexDatabaseWriterTest {
 
-    private StringWriter stringWriter;
     private BibtexDatabaseWriter databaseWriter;
     private BibDatabase database;
     private MetaData metaData;
@@ -70,17 +69,20 @@ class BibtexDatabaseWriterTest {
     private GeneralPreferences generalPreferences;
     private SavePreferences savePreferences;
     private BibEntryTypesManager entryTypesManager;
+    private StringWriter stringWriter;
+    private BibWriter bibWriter;
 
     @BeforeEach
     void setUp() {
-        stringWriter = new StringWriter();
         generalPreferences = mock(GeneralPreferences.class);
         when(generalPreferences.getDefaultEncoding()).thenReturn(null);
         savePreferences = mock(SavePreferences.class, Answers.RETURNS_DEEP_STUBS);
         when(savePreferences.getSaveOrder()).thenReturn(new SaveOrderConfig());
         when(savePreferences.takeMetadataSaveOrderInAccount()).thenReturn(true);
         entryTypesManager = new BibEntryTypesManager();
-        databaseWriter = new BibtexDatabaseWriter(stringWriter, generalPreferences, savePreferences, entryTypesManager);
+        stringWriter = new StringWriter();
+        bibWriter = new BibWriter(stringWriter, OS.NEWLINE);
+        databaseWriter = new BibtexDatabaseWriter(bibWriter, generalPreferences, savePreferences, entryTypesManager);
 
         database = new BibDatabase();
         metaData = new MetaData();
@@ -113,7 +115,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Preamble{Test preamble}" + OS.NEWLINE, stringWriter.toString());
+        assertEquals("@Preamble{Test preamble}" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -135,9 +137,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.singletonList(entry));
 
-        assertEquals(
-                OS.NEWLINE +
-                        "@Article{," + OS.NEWLINE + "}"
+        assertEquals("@Article{," + OS.NEWLINE + "}"
                         + OS.NEWLINE,
                 stringWriter.toString());
     }
@@ -164,7 +164,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "Test epilog" + OS.NEWLINE, stringWriter.toString());
+        assertEquals("Test epilog" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -186,7 +186,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Comment{jabref-meta: keypatterndefault:test;}" + OS.NEWLINE,
+        assertEquals("@Comment{jabref-meta: keypatterndefault:test;}" + OS.NEWLINE,
                 stringWriter.toString());
     }
 
@@ -213,8 +213,7 @@ class BibtexDatabaseWriterTest {
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
         // @formatter:off
-        assertEquals(OS.NEWLINE
-                + "@Comment{jabref-meta: grouping:" + OS.NEWLINE
+        assertEquals("@Comment{jabref-meta: grouping:" + OS.NEWLINE
                 + "0 AllEntriesGroup:;" + OS.NEWLINE
                 + "1 StaticGroup:test\\;2\\;1\\;\\;\\;\\;;" + OS.NEWLINE
                 + "}" + OS.NEWLINE, stringWriter.toString());
@@ -248,7 +247,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
+        assertEquals("@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -260,6 +259,16 @@ class BibtexDatabaseWriterTest {
 
         assertEquals("% Encoding: US-ASCII" + OS.NEWLINE + OS.NEWLINE +
                 "@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    void doNotWriteUtf8StringAndEncoding() throws Exception {
+        when(generalPreferences.getDefaultEncoding()).thenReturn(StandardCharsets.UTF_8);
+        database.addString(new BibtexString("name", "content"));
+
+        databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
+
+        assertEquals("@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -285,9 +294,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.saveDatabase(bibtexContext);
 
-        assertEquals(
-                OS.NEWLINE +
-                        "@Customizedtype{," + OS.NEWLINE + "}" + OS.NEWLINE + OS.NEWLINE
+        assertEquals("@Customizedtype{," + OS.NEWLINE + "}" + OS.NEWLINE + OS.NEWLINE
                         + "@Comment{jabref-meta: databaseType:bibtex;}"
                         + OS.NEWLINE + OS.NEWLINE
                         + "@Comment{jabref-entrytype: customizedtype: req[title;author;date] opt[year;month;publisher]}" + OS.NEWLINE,
@@ -317,8 +324,7 @@ class BibtexDatabaseWriterTest {
         databaseWriter.savePartOfDatabase(bibtexContext, Arrays.asList(entry, otherEntry));
 
         assertEquals(
-                OS.NEWLINE
-                        + "@Customizedtype{," + OS.NEWLINE + "}" + OS.NEWLINE + OS.NEWLINE
+                "@Customizedtype{," + OS.NEWLINE + "}" + OS.NEWLINE + OS.NEWLINE
                         + "@Othercustomizedtype{," + OS.NEWLINE + "}" + OS.NEWLINE + OS.NEWLINE
                         + "@Comment{jabref-meta: databaseType:bibtex;}"
                         + OS.NEWLINE + OS.NEWLINE
@@ -339,6 +345,33 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(context, result.getDatabase().getEntries());
         assertEquals(Files.readString(testBibtexFile, encoding), stringWriter.toString());
+    }
+
+    @Test
+    void roundtripUtf8EncodingHeaderRemoved() throws Exception {
+        when(generalPreferences.getDefaultEncoding()).thenReturn(StandardCharsets.UTF_8);
+        // @formatter:off
+        String bibtexEntry = OS.NEWLINE + "% Encoding: UTF8" + OS.NEWLINE +
+                OS.NEWLINE +
+                "@Article{," + OS.NEWLINE +
+                "  author  = {Foo Bar}," + OS.NEWLINE +
+                "  journal = {International Journal of Something}," + OS.NEWLINE +
+                "  note    = {some note}," + OS.NEWLINE +
+                "  number  = {1}," + OS.NEWLINE +
+                "}" + OS.NEWLINE;
+        // @formatter:on
+        ParserResult result = new BibtexParser(importFormatPreferences, fileMonitor).parse(new StringReader(bibtexEntry));
+        BibDatabaseContext context = new BibDatabaseContext(result.getDatabase(), result.getMetaData());
+        databaseWriter.saveDatabase(context);
+        // @formatter:off
+        String expected = "@Article{," + OS.NEWLINE +
+                "  author  = {Foo Bar}," + OS.NEWLINE +
+                "  journal = {International Journal of Something}," + OS.NEWLINE +
+                "  note    = {some note}," + OS.NEWLINE +
+                "  number  = {1}," + OS.NEWLINE +
+                "}" + OS.NEWLINE;
+        // @formatter:on
+       assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -367,6 +400,64 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(context, result.getDatabase().getEntries());
         assertEquals(Files.readString(testBibtexFile, encoding), stringWriter.toString());
+    }
+
+    @Test
+    void roundtripWithOneUserCommentAndEntryChange() throws Exception {
+        String bibEntry = "@Comment this in an unbracketed comment that should be preserved as well\n" +
+                "\n" +
+                "This is some arbitrary user comment that should be preserved\n" +
+                "\n" +
+                "@InProceedings{1137631,\n" +
+                "  author     = {Mr. Author},\n" +
+                "}\n";
+
+        // read in bibtex string
+        ImportFormatPreferences importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
+        ParserResult result = new BibtexParser(importFormatPreferences, new DummyFileUpdateMonitor()).parse(new StringReader(bibEntry));
+
+        BibEntry entry = result.getDatabase().getEntryByCitationKey("1137631").get();
+        entry.setField(StandardField.AUTHOR, "Mr. Author");
+
+        when(generalPreferences.getDefaultEncoding()).thenReturn(StandardCharsets.UTF_8);
+        when(savePreferences.shouldSaveInOriginalOrder()).thenReturn(true);
+        BibDatabaseContext context = new BibDatabaseContext(result.getDatabase(), result.getMetaData());
+
+        // we need a new writer because "\n" instead of "OS.NEWLINE"
+        bibWriter = new BibWriter(stringWriter, "\n");
+        databaseWriter = new BibtexDatabaseWriter(bibWriter, generalPreferences, savePreferences, entryTypesManager);
+        databaseWriter.savePartOfDatabase(context, result.getDatabase().getEntries());
+        assertEquals(bibEntry, stringWriter.toString());
+    }
+
+    @Test
+    void roundtripWithTwoEntriesAndOneUserCommentAndEntryChange() throws Exception {
+        String bibEntry = "@Article{test,}\n" +
+                "\n" +
+                "@Comment this in an unbracketed comment that should be preserved as well\n" +
+                "\n" +
+                "This is some arbitrary user comment that should be preserved\n" +
+                "\n" +
+                "@InProceedings{1137631,\n" +
+                "  author     = {Mr. Author},\n" +
+                "}\n";
+
+        // read in bibtex string
+        ImportFormatPreferences importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
+        ParserResult result = new BibtexParser(importFormatPreferences, new DummyFileUpdateMonitor()).parse(new StringReader(bibEntry));
+
+        BibEntry entry = result.getDatabase().getEntryByCitationKey("1137631").get();
+        entry.setField(StandardField.AUTHOR, "Mr. Author");
+
+        when(generalPreferences.getDefaultEncoding()).thenReturn(StandardCharsets.UTF_8);
+        when(savePreferences.shouldSaveInOriginalOrder()).thenReturn(true);
+        BibDatabaseContext context = new BibDatabaseContext(result.getDatabase(), result.getMetaData());
+
+        // we need a new writer because "\n" instead of "OS.NEWLINE"
+        bibWriter = new BibWriter(stringWriter, "\n");
+        databaseWriter = new BibtexDatabaseWriter(bibWriter, generalPreferences, savePreferences, entryTypesManager);
+        databaseWriter.savePartOfDatabase(context, result.getDatabase().getEntries());
+        assertEquals(bibEntry, stringWriter.toString());
     }
 
     @Test
@@ -431,7 +522,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.singletonList(entry));
 
-        assertEquals("presaved serialization", stringWriter.toString());
+        assertEquals("presaved serialization" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -446,9 +537,7 @@ class BibtexDatabaseWriterTest {
         when(savePreferences.shouldReformatFile()).thenReturn(true);
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.singletonList(entry));
 
-        assertEquals(
-                OS.NEWLINE
-                        + "@Article{," + OS.NEWLINE + "  author = {Mr. author}," + OS.NEWLINE + "}"
+        assertEquals("@Article{," + OS.NEWLINE + "  author = {Mr. author}," + OS.NEWLINE + "}"
                         + OS.NEWLINE,
                 stringWriter.toString());
     }
@@ -461,7 +550,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals("serialization", stringWriter.toString());
+        assertEquals("serialization" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -473,7 +562,7 @@ class BibtexDatabaseWriterTest {
         when(savePreferences.shouldReformatFile()).thenReturn(true);
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
+        assertEquals("@String{name = {content}}" + OS.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -487,9 +576,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(
-                OS.NEWLINE +
-                        "@Comment{jabref-meta: saveActions:enabled;"
+        assertEquals("@Comment{jabref-meta: saveActions:enabled;"
                         + OS.NEWLINE
                         + "day[upper_case]" + OS.NEWLINE
                         + "journal[title_case]" + OS.NEWLINE
@@ -508,8 +595,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE
-                + "@Comment{jabref-meta: saveOrderConfig:specified;author;false;year;true;abstract;false;}"
+        assertEquals("@Comment{jabref-meta: saveOrderConfig:specified;author;false;year;true;abstract;false;}"
                 + OS.NEWLINE, stringWriter.toString());
     }
 
@@ -522,7 +608,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Comment{jabref-meta: keypattern_article:articleTest;}" + OS.NEWLINE
+        assertEquals("@Comment{jabref-meta: keypattern_article:articleTest;}" + OS.NEWLINE
                         + OS.NEWLINE + "@Comment{jabref-meta: keypatterndefault:test;}" + OS.NEWLINE,
                 stringWriter.toString());
     }
@@ -533,7 +619,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Comment{jabref-meta: databaseType:biblatex;}" + OS.NEWLINE,
+        assertEquals("@Comment{jabref-meta: databaseType:biblatex;}" + OS.NEWLINE,
                 stringWriter.toString());
     }
 
@@ -543,7 +629,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Comment{jabref-meta: protectedFlag:true;}" + OS.NEWLINE,
+        assertEquals("@Comment{jabref-meta: protectedFlag:true;}" + OS.NEWLINE,
                 stringWriter.toString());
     }
 
@@ -555,7 +641,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, Collections.emptyList());
 
-        assertEquals(OS.NEWLINE + "@Comment{jabref-meta: fileDirectory:\\\\Literature\\\\;}" + OS.NEWLINE +
+        assertEquals("@Comment{jabref-meta: fileDirectory:\\\\Literature\\\\;}" + OS.NEWLINE +
                 OS.NEWLINE + "@Comment{jabref-meta: fileDirectory-defaultOwner-user:D:\\\\Documents;}"
                 + OS.NEWLINE + OS.NEWLINE + "@Comment{jabref-meta: fileDirectoryLatex-defaultOwner-user:D:\\\\Latex;}" + OS.NEWLINE, stringWriter.toString());
     }
@@ -589,9 +675,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(bibtexContext, database.getEntries());
 
-        assertEquals(
-                OS.NEWLINE +
-                        "@Article{," + OS.NEWLINE +
+        assertEquals("@Article{," + OS.NEWLINE +
                         "  author = {A}," + OS.NEWLINE +
                         "  year   = {2010}," + OS.NEWLINE +
                         "}" + OS.NEWLINE + OS.NEWLINE +
@@ -632,9 +716,7 @@ class BibtexDatabaseWriterTest {
         when(savePreferences.shouldSaveInOriginalOrder()).thenReturn(false);
         databaseWriter.savePartOfDatabase(bibtexContext, database.getEntries());
 
-        assertEquals(
-                OS.NEWLINE +
-                        "@Article{," + OS.NEWLINE +
+        assertEquals("@Article{," + OS.NEWLINE +
                         "  author = {A}," + OS.NEWLINE +
                         "  year   = {2010}," + OS.NEWLINE +
                         "}" + OS.NEWLINE + OS.NEWLINE +
@@ -658,8 +740,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.saveDatabase(bibtexContext);
 
-        assertEquals(
-                OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        assertEquals("@Article{," + OS.NEWLINE +
                         "  note = {some note}," + OS.NEWLINE +
                         "}" + OS.NEWLINE,
                 stringWriter.toString());
@@ -675,8 +756,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.saveDatabase(bibtexContext);
 
-        assertEquals(
-                OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        assertEquals("@Article{," + OS.NEWLINE +
                         "  abstract = {" + text + "}," + OS.NEWLINE +
                         "}" + OS.NEWLINE,
                 stringWriter.toString());
@@ -684,7 +764,9 @@ class BibtexDatabaseWriterTest {
 
     @Test
     void roundtripWithContentSelectorsAndUmlauts() throws Exception {
-        String fileContent = "% Encoding: UTF-8" + OS.NEWLINE + OS.NEWLINE + "@Comment{jabref-meta: selector_journal:Test {\\\\\"U}mlaut;}" + OS.NEWLINE;
+        String encodingHeader = "% Encoding: UTF-8" + OS.NEWLINE + OS.NEWLINE;
+        String commentEntry = "@Comment{jabref-meta: selector_journal:Test {\\\\\"U}mlaut;}" + OS.NEWLINE;
+        String fileContent = encodingHeader + commentEntry;
         Charset encoding = StandardCharsets.UTF_8;
 
         ParserResult firstParse = new BibtexParser(importFormatPreferences, fileMonitor).parse(new StringReader(fileContent));
@@ -695,7 +777,7 @@ class BibtexDatabaseWriterTest {
 
         databaseWriter.savePartOfDatabase(context, firstParse.getDatabase().getEntries());
 
-        assertEquals(fileContent, stringWriter.toString());
+        assertEquals(commentEntry, stringWriter.toString());
     }
 
     @Test
@@ -728,11 +810,11 @@ class BibtexDatabaseWriterTest {
 
         // write a second time
         stringWriter = new StringWriter();
-        databaseWriter = new BibtexDatabaseWriter(stringWriter, generalPreferences, savePreferences, entryTypesManager);
+        bibWriter = new BibWriter(stringWriter, OS.NEWLINE);
+        databaseWriter = new BibtexDatabaseWriter(bibWriter, generalPreferences, savePreferences, entryTypesManager);
         databaseWriter.savePartOfDatabase(context, firstParse.getDatabase().getEntries());
 
-        assertEquals(OS.NEWLINE +
-                "@Article{test," + OS.NEWLINE +
+        assertEquals("@Article{test," + OS.NEWLINE +
                 "  author  = {Test}," + OS.NEWLINE +
                 "  journal = {International Journal of Something}," + OS.NEWLINE +
                 "  note    = {some note}," + OS.NEWLINE +
@@ -745,12 +827,12 @@ class BibtexDatabaseWriterTest {
     @Test
     void saveReturnsToOriginalEntryWhenEntryIsFlaggedUnchanged() throws Exception {
         // @formatter:off
-        String bibtexEntry = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
+        String bibtexEntry = "@Article{test," + OS.NEWLINE +
                 "  Author                   = {Foo Bar}," + OS.NEWLINE +
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Number                   = {1}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -766,8 +848,6 @@ class BibtexDatabaseWriterTest {
         entry.setChanged(false);
 
         // write entry
-        stringWriter = new StringWriter();
-        databaseWriter = new BibtexDatabaseWriter(stringWriter, generalPreferences, savePreferences, entryTypesManager);
         BibDatabaseContext context = new BibDatabaseContext(firstParse.getDatabase(), firstParse.getMetaData());
         databaseWriter.savePartOfDatabase(context, firstParse.getDatabase().getEntries());
 
@@ -775,14 +855,14 @@ class BibtexDatabaseWriterTest {
     }
 
     @Test
-    void saveReturnsToOriginalEntryWhenEntryIsFlaggedUnchangedEvenInThePrecenseOfSavedModifications() throws Exception {
+    void saveReturnsToOriginalEntryWhenEntryIsFlaggedUnchangedEvenInThePresenceOfSavedModifications() throws Exception {
         // @formatter:off
-        String bibtexEntry = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
+        String bibtexEntry = "@Article{test," + OS.NEWLINE +
                 "  Author                   = {Foo Bar}," + OS.NEWLINE +
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
                 "  Number                   = {1}," + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -805,7 +885,8 @@ class BibtexDatabaseWriterTest {
 
         // write a second time
         stringWriter = new StringWriter();
-        databaseWriter = new BibtexDatabaseWriter(stringWriter, generalPreferences, savePreferences, entryTypesManager);
+        bibWriter = new BibWriter(stringWriter, OS.NEWLINE);
+        databaseWriter = new BibtexDatabaseWriter(bibWriter, generalPreferences, savePreferences, entryTypesManager);
         databaseWriter.savePartOfDatabase(context, firstParse.getDatabase().getEntries());
 
         // returns tu original entry, not to the last saved one